Sweep mode sets header once per group of blocks
Before this change, in sweep mode, every block of 8192 samples had its
first 5 samples replaced with a header containing information about the
current center frequency. This change removes the header for all blocks
but the first one in every group. Besides, if the first sample in the
second or any subsequent block matches the header marker it is modified
by decrementing the I component.